pyproject.toml
[project]
name = "protein-embedder"
version = "0.1.0"
description = "A Python library for processing protein FASTA files and generating embeddings using pre-trained models."
authors = [
    { name = "Chaima Ben Farhat" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "biopython>=1.79",
    "torch>=1.9.0",
    "transformers>=4.20.0",
    "numpy>=1.21.0",
    "tqdm>=4.62.0"
]
keywords = ["protein", "embedding", "bioinformatics", "machine learning"]

[project.urls]
Homepage = ""
Repository = ""
Issues = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/protein_embedder"]